+++ newIDE/app/src/UI/CustomSvgIcons/ChevronArrowRightWithRoundedBorder.js
