--- newIDE/app/src/UI/CustomSvgIcons/SquaredDoubleChevronArrowDown.js
@@ -1,28 +1,0 @@
-import React from 'react';
-import SvgIcon from '@material-ui/core/SvgIcon';
-
-export default React.memo(props => (
-  <SvgIcon {...props} width="16" height="16" viewBox="0 0 16 16" fill="none">
-    <rect
-      x="0.5"
-      y="0.5"
-      width="15"
-      height="15"
-      rx="3.5"
-      stroke="currentColor"
-      fill="none"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M4.93304 5.03353C5.09872 4.81261 5.41212 4.76784 5.63304 4.93353L7.9997 6.70853L10.3664 4.93353C10.5873 4.76784 10.9007 4.81261 11.0664 5.03353C11.2321 5.25444 11.1873 5.56784 10.9664 5.73353L8.2997 7.73353C8.12193 7.86686 7.87748 7.86686 7.6997 7.73353L5.03304 5.73353C4.81212 5.56784 4.76735 5.25444 4.93304 5.03353Z"
-      fill="currentColor"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M4.93304 8.36686C5.09872 8.14595 5.41212 8.10117 5.63304 8.26686L7.9997 10.0419L10.3664 8.26686C10.5873 8.10117 10.9007 8.14595 11.0664 8.36686C11.2321 8.58777 11.1873 8.90117 10.9664 9.06686L8.2997 11.0669C8.12193 11.2002 7.87748 11.2002 7.6997 11.0669L5.03304 9.06686C4.81212 8.90117 4.76735 8.58777 4.93304 8.36686Z"
-      fill="currentColor"
-    />
-  </SvgIcon>
-));
