--- newIDE/app/src/UI/CustomSvgIcons/SquaredDoubleChevronArrowUp.js
@@ -1,28 +1,0 @@
-import React from 'react';
-import SvgIcon from '@material-ui/core/SvgIcon';
-
-export default React.memo(props => (
-  <SvgIcon {...props} width="16" height="16" viewBox="0 0 16 16" fill="none">
-    <rect
-      x="0.5"
-      y="0.5"
-      width="15"
-      height="15"
-      rx="3.5"
-      stroke="currentColor"
-      fill="none"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M4.93304 7.6335C5.09872 7.85441 5.41212 7.89918 5.63304 7.7335L7.9997 5.9585L10.3664 7.7335C10.5873 7.89918 10.9007 7.85441 11.0664 7.6335C11.2321 7.41258 11.1873 7.09918 10.9664 6.9335L8.2997 4.9335C8.12193 4.80016 7.87748 4.80016 7.6997 4.9335L5.03304 6.9335C4.81212 7.09918 4.76735 7.41258 4.93304 7.6335Z"
-      fill="currentColor"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M4.93304 10.9668C5.09872 11.1877 5.41212 11.2325 5.63304 11.0668L7.9997 9.29183L10.3664 11.0668C10.5873 11.2325 10.9007 11.1877 11.0664 10.9668C11.2321 10.7459 11.1873 10.4325 10.9664 10.2668L8.2997 8.26683C8.12193 8.1335 7.87748 8.1335 7.6997 8.26683L5.03304 10.2668C4.81212 10.4325 4.76735 10.7459 4.93304 10.9668Z"
-      fill="currentColor"
-    />
-  </SvgIcon>
-));
